Make some small changes and updated comments.

+++ get-beats.py
@@ -1,9 +1,16 @@
+# Small helper file to download the latest beats
+# May need to change the version number and the beats collection
+
 import urllib.request
 import os
 
+# Create the beats directory
 directory = ".\\beats"
 if not os.path.exists(directory):
+    print("Creating 'beats' directory")
     os.makedirs(directory)
+else:
+    print("'beats' directory already exists")
 
 # All the beat types that will be processed
 beats = ['winlogbeat',
@@ -11,8 +18,14 @@ beats = ['winlogbeat',
          #'metricbeat', # not using based on inital hunt slides
          'packetbeat']
 
+# Get the beats
 for beat in beats:
     version = "6.0.0"
     file_name = beat + "-"+ version +"-windows-x86_64.zip"
-    url = "https://artifacts.elastic.co/downloads/beats/" + beat + "/" + file_name
-    urllib.request.urlretrieve(url, directory + "\\" + file_name)
+    if not os.path.exists(directory + "\\" + file_name):
+        url = "https://artifacts.elastic.co/downloads/beats/" + beat + "/" + file_name
+        print(f"\tGetting {beat} from: {url}", end='')
+        urllib.request.urlretrieve(url, directory + "\\" + file_name)
+        print("...complete")
+    else:
+        print(f"\tFile {file_name} already exists")
